requirements.txt
psutil>=5.9
pyyaml>=6.0
winotify>=1.1;sys_platform=="win32"
watchdog>=3.0
